[workloadmeta] Move proto utils to workloadmeta dir
