+++ comp/core/workloadmeta/proto/proto.go
@@ -3,6 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2020 Datadog, Inc.
 
+// Package proto provides conversions between Workloadmeta types and protobuf.
 package proto
 
 import (
